requirements.txt
# Core Data Processing
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.1
dask

# Azure Services
azure-identity>=1.13.0
azure-storage-blob>=12.14.0
azure-storage-file-datalake>=12.12.0
azure-functions>=1.14.0
azure-mgmt-datafactory>=2.0.0
azure-mgmt-resource>=21.1.0
azure-mgmt-storage>=20.0.0
azure-monitor-query>=1.2.0
azure-keyvault-secrets>=4.6.0
databricks-sdk>=0.1.0

# ML & Data Quality
mlflow>=2.4.1
great-expectations>=0.13.44
ydata-profiling>=4.6.0

# Dashboard & Visualization
streamlit>=1.24.0
plotly>=5.15.0

# Development & Testing
python-dotenv>=1.0.0
pytest>=6.2.5

# Note: The following packages should be installed via conda if needed:
# - apache-spark
# - delta-spark
